Add the bd_lvm_pvinfo function and all it needs
The bd_lvm_pvinfo function needs to return data about a given PV device. Since
we want to support GObject introspection, there are three choices:
1) an opaque structure with accessor methods for every item
2) a GObject-inherited class with all the boileplate code it needs
3) a bare C struct boxed as a minimalistic GType

Number 3 requires least code to be written and provides nice functionality and
syntax for both C and non-C code using it. The only requirements are the
function for freeing the struct and the function for creating a copy so that the
introspection layers in non-C languages know how to duplicate and delete such
structs.