Generate both .c and .h files for plugins
Instead of generating .c files from plugins' .h files, it is better to generate
both .c and .h files from a file that is used only as a source of data for
generating those files. This way we have a better control of what ends up in the
.h file and what ends up in the .c file. In particular we can have .h files that
contain no code and can thus be easily included by other code.

For the build process it is also better to have source files for the generator
using a unique extension because that simplifies pattern matching.