Fix some minor issues in the spec file
The %{_includedir}/blockdev directory has to be owned by some package and since
installing any of our -devel packages on its own makes perfect sense, all these
packages should own the directory so that it is not unowned in any case.

Also the dependency on the libblockdev-utils-devel package should specify
architecture.