Implement the ErrorProxy python class and use it
The GLib.Error class is not really nice to work for and with all the libblockdev
calls from python raising instances of the same exception class it's hard to do
any nice further processing of such exception. However, we can provide an error
proxy that makes sure that the GLib.Error instances are transformed into native
Python exceptions with a nice hierarchy and better granularity thanks to having
an instance of the ErrorProxy class for every plugin libblockdev supports.