Put the GI overrides in a special dir so that they are preferred
Having the GI overrides file (BlockDev.py) under gi/overrides directory
structure rooted at something that is on PYTHONPATH does the trick for the
local overrides to be used instead of the installed version (or none if not
installed). This way we can finally test the overrides themselves locally
without the need of installing everything to the system. Yay!