Raise error/exception when an unimplemented function is called (#1201475)
Printing out some message to stderr is not "loud enough" even if it is CRITICAL
and can be silently ignored (especially if the stderr is redirected somewhere).
We need to report error/raise exception if an unimplemented function is called.
Such error/exception can be handled by the caller just fine or they can
propagate it if they want to.