Explicitly specify the type of the cert_data parameter (#1246096)
GObject introspection translates 'gchar*' into 'utf8', but the cert_data
parameter holds binary data which are not UTF-8 encoded and which are usually
passed from the outside (like in Python as described in the bug) as binary data
and not as a string. On the other hand NSS expects certificate data as a
character buffer so we need to just take a character buffer from the caller
(instead of doing some transformation magic). Thus we need to specify the type
explicitly to a zero-terminated array of chars which resolves the issue on the
caller interface and keeps everything okay for NSS.