Fix the calculation of physical space taken by an LV
If an LV is added to a VG, no extra meta data is allocated and it doesn't matter
at all if one or more of the VG's PVs are MD RAID devices or whatever, they are
just block devices. In the end we just need to make sure the LV size is rounded
up to the PE size.

Also update the tests accordingly.