Add functions for finding and processing configuration files
In order to support multiple plugins per technology (i.e. multiple
implementations of the plugins APIs), we need to let plugins provide a
configuration file that specifies priorities of their implementations (so that
the more optimized ones are preferred) and thus we need to be able to find and
process such configuration files in the right order.