Load and respect configuration files when loading plugins
Configuration plugins (may) give us the priority and fallback rules for plugins'
implementations. Thus when loading plugins we should load and respect such
configuration by loading plugins' implementations (shared objects) in the
specified order.

However, if some particular implementations are required/requested by the user
code as part of the call to one of the *init() functions, we should still try to
load only the given implementation. In case it is needed in future the logic
could be extended so that the new most preferred or fallback implementations
could be given in the *init() calls.