Add a function for getting the loaded soname for a plugin
With the ability to try loading mutliple shared objects for a plugin in the
fallback fashion we need to add a way for the user code to find out which of the
shared objects was loaded in the end (if any). This requires us to remember the
name of the successfully loaded shared object.

Also remove traces of a planned, but never implemented function
bd_func_available.