Use addCleanup() instead of tearDown() in tests
Python's unittests have a weird glitch -- the tearDown() method of a test case
is not called always. There are cases (exception in setUp(), exception in the
test?,...) when it's just skipped. However, we need to always clean after
ourselves and that's why need to use the addCleanup() method which adds
callbacks that are called ALWAYS when the test is done (succeeding, failing,
erroring, whatever).