Better check for the LVM DBus API
The fact that the service is running doens't automatically mean that we can
access and use it. In order to check for that we may simply try to introspect
the root node which is guaranteed to always be there. If things work nicely, we
may also trigger autostart the service this way.