Rework how we do optimal alignment
It seems that when somebody tries to create small partitions starting close to
the start of the disk with optimal alignment, cylinder alignment can come into
the game the result of which is e.g. a 265MiB partition created instead of 2MiB
one. Thus we need to turn the cylinder alignment off when adding partitions and
then reset it later to prevent some unwanted changes.

It's also much better to create a geometry and let it do the calculations from
(start, size) to (start, end).

And since the optimal alignment now works properly, we need to change the tests
that relied on the old behaviour.

This fixes the issue #82.

(cherry picked from commit 0acd5b751d7cc4219f8997104ce4aed09fe2ca50)