Use different BTRFS_MIN_MEMBER_SIZE on aarch64
It seems to be 256 MiB instead of 128 MiB (don't ask me why). Originally I used
'#ifdef __aarch64__' in the sources, but guess what? GI doesn't really get that
and so the latter value was always used no matter what was the architecture. To
work around this the value is actually set in configure.ac based on the output
of the 'uname -p' command.