Add a function for setting multiple partition flags at once
Not sure how often multiple partition flags are set, but if it happens, it's
better to do it in a single function call with a single commit to disk. This
also brings some atomicity to the operation as if (un)setting one flag fails,
nothing is commited to the disk.

Requested in the issue #81.