Add function for getting partition by position
A similar thing is done in Blivet in multiple places which proves it can be
handy for other storage configuration tools as well.

The tests revealed an issue with BDPartSpec.type being of type BDPartType which
didn't support representing free regions on extended partitions (which are both
LOGICAL and FREESPACE). It needs to be an unsigned 64bit integer serving as a
bitwise OR combination of the types.

(cherry picked from commit 9aff7f7f9066ba524939b369939ab3c96ae0454b)