Add progress reporting to the LVM plugins
The only operation that LVM CLI provides some progress for is 'pvmove' which we
already have covered (the recent proof-of-concept). For other operations we have
no idea so we can just report start and finish which is done by our exec*
functions being used. The only exception are two cache-related functions which
perform multiple steps/operations allowing us to report progress based on that.

The DBus API may or may not provide progress for any operation (method call) and
we are prepared to handle that. So again nothing to do here except for the two
cache-related functions which can use the same trick as mentioned above.