Do not try to get progress from LVM DBus if there was a failure
If we fail to get the state of the 'Complete' property of a Job object, it
doesn't make sense to try to get the state of another property from it. This can
only lead to issues.

Also fix the variant format string for object path results. It seems to finally
be a real object path instead of just string.