Add a way for plugins to clean after themselves
Plugins can define their check() and init() functions to check if all the
resources they need are available and initialize them if needed. To prevent
resource leakage when the plugins are unloaded they need a way to clean after
themselves - a close_plugin() function.

Note: The function is not called close() because that function seems to be
      defined for every library.