Try to search for "RAID Level" in mdadm's output (#1379865)
mdadm's API (CLI) is ambiguous and may provide the RAID level as either the
"RAID Level" or "Raid Level" item. Since we store these items in a hashtable,
let's try to lookup both if the first one fails. Another approach would be to
cannonicalize all the items keys when creating the hash table, but that seems
like an overkill for one weird quirk.