Change how we run tests
Having a universal target for running only tests for a single plugins is
extremely useful during development. And we can use this target to run tests for
all plugins one after another. That gives a better idea of how long each set of
tests runs and where the problematic areas are. And it should allow us running
tests in parallel (maybe on multiple nodes?) in the future.