Add extra parameters to loop_setup required by storaged
storaged supports almost everything that's possible to do as part of setting up
a loop device. In order to be a sufficient backend for storaged we need to
support the same set of features.

We need to reorganize the tests a little bit to make sure that everything is
cleaned up properly in case of failure.