Prevent issues between libparted and udev
libparted does committing of a new partition table to disk in two steps:

1. committing (writing and flushing) to disk and
2. telling OS about the partitions.

However, that means that if things go wrong, udev may step in between the above
steps and trigger events when the block device file is closed (from being opened
RW) that could lead to a failure of the second step because some other process
may open the block device file in the reaction to the event.

In order to prevent such issues we need to acquire an exclusive lock on the
block device file, which prevents udev from triggering events, for the time we
let libparted perform the two operations above.

(cherry picked from commit 0a65466154994759cbec15ce984b751b073a891f)