Do not require an exclusive lock on the device to commit part stuff
Something else (e.g. the caller) may be holding a shared lock on the device
which means libblockdev cannot get an exclusive one. A shared lock should be
okay as it is enough for preventing udev from calling the BLKRRPART ioctl() on
the device which is the major source of issues. There's still a possibility that
something else will step in because with a shared lock udev triggers events on
the device, but that has very small chance of affecting us. And we don't need to
open the file as RDWR because we are not blocking udev from triggering events.

Also, don't strictly require the lock. If we fail to get it, we can just move on
hoping for the best, there will still be chances that things will work just fine
(quite high, actually, because not getting the lock means somebody's holding an
exclusive lock on the device preventing udev from doing anything with it).