Deal with older parted and libblkid
Older versions of parted don't provide a separate pkg-config package for the
libparted-fs-resize library so we must define the necessary stuff on our own in
such cases.

Older versions of libblkid don't support the BLKID_SUBLKS_BADCSUM flag so again
we must define it on our own to prevent compilation errors. Using a neutral
value 0 means our definition of the flag doesn't contribute to bit-combinations
of flags and has thus no functional effect. Which is exactly what we want with
older versions of libblkid.