Make sure bcache functions are correctly provided or not
When compiling libblockdev, we define macros to make sure bcache support is
compiled in or not. However, when the library is built and we ship the header
files, there's no way to determine if the bcache support is built in or not and
the macros are of course no longer defined. By providing cflags for compiling
things with libblockdev, we can propagate the information if bcache is supported
or not.