Try harder to register a new bcache device
Sometimes the /sys/.../register file doesn't exist when we try to write to
it. Let's give it a chance to appear in some sensible time frame. With some luck
the file appears and the bcache device is successfully registered (and then
properly created and set up).

Also enable some of the risky tests that seem to have been failing due to the
issue fixed here.