Resolve maximum partition size when we know its start
Otherwise the requested size may be too big if the start moves
due to alignment. Also we need to subtract start from end to get
the actual size. Some older versions of libparted are not so
benevolent in this and just return an error instead of adapting
the geometry to the device's geometry.

Also fix a nonsense in the tests -- next partition cannot start
where the previous one starts (and again older version of
libparted don't fix our stupidity).