Make sure the loop device doesn't disappear during tests
Once we set 'autoclear' to 'True', the loop device being used for
tests may be cleared automatically by kernel because it is not
being used (not open, mounted,...). But since we want to use it
for further testing (e.g. setting 'autoclear' back to 'False') we
need to make sure it stays in place. Having it open for reading
is enough so let's just do that.