Ignore previous errors when falling back to using ioctl()
If we cannot get the 'autoclear' flag from /sys we need to give
it another try using the ioctl() call, but in no case we want to
report the error that may have potentially been set when trying
to read the value from /sys. We either want to set a new error
(which shouldn't be set over an old one) or report no error.