Use a special exception for no fs signature when doing wipe
If bd_fs_wipe() is called on a device where there's nothing to
wipe (i.e. no signature) an error/exception with a special code
and message is raised. There should be a special python exception
for this case in Python too, otherwise the user code has to catch
all the FSError exceptions and check their messages.