kbd.c: Make bd_kbd_bcache_create work without abort
When you create a cached device with a backing device at the same
time you do not need to separately register and attach devices.
This is done by the udev rules.  This fixes an issue where the
code was doing an attach for a device that was already attached.
The bcache sysfs attach kernel entry will return -2 in an unsigned,
which yields 2**32-2 returned to user space.  This will cause an
abort if you are running G_DEBUG=fatal-criticals in glib as
g_io_channel_flush does a g_string_erase of the write buffer for
the number of bytes written, which is ~4G, and we only actually
have a 1K buffer that we wrote 36 bytes of.

Signed-off-by: Tony Asleson <tasleson@redhat.com>