Add partition resize function
The new function bd_part_resize_part allows to change the size
of a partition and guarantees that it does not move and has at
least the requested size allocated. It also supports resizing
to maximum in a reliable way through the libparted
ped_disk_get_max_partition_geometry.
The code is shared with add_part_to_disk and also resolves its
bug related adding a partition in front of another one while it
also should occupy the maximal space. This is achieved by
creating a dummy partition of 1 MiB and call the resize on it
because ped_disk_get_max_partition_geometry can only be used
for already existing partitions.