Do not strictly require all FS utilities
We now have code for checking if FS operations are
supported/available or not. Having the plugin to fail to load
when not everything it could possibly need is available is
pointless. We need a much better way to do the checks in plugins
in general. But for now let's just make the FS plugin usable.

Related: #217