Do not verify vfat FS' size after generic resize
Getting information about the file system after the resize
sometimes fails. Totally randomly. Which is a pity, but we
definitely don't want our CI results to become almost useless due
to this.

Also, other file systems do the resize precisely so we don't have
to be benevolent when comparing the requested and resulting
sizes. Which means better tests for the other file systems.