Make sure the device is opened for libparted
Sometimes it magically happens that libparted kills the whole
process trying to create a partition due to the

  assert(disk->dev->open_count > 0)

Which should obviously never happen because we are never opening
or closing the device explicitly just by means of
ped_device_get() and ped_device_destroy(). So this is probably a
bug in libparted, but we have no reproducer for it.

Let's just try to make sure the device is opened while we are
trying to commit the new partition table to it.