Dynamically check for the required utilities
If LIBBLOCKDEV_SKIP_DEP_CHECKS is used and we skip the dependency
checks when loading plugins (so that the plugins load even with
something is missing), we need to check the availability of the
required utilities once they are needed. Otherwise cryptic errors
are given (like "Failed to do SOMETHING: No such file or
directory") and functions fail in the middle of their jobs.

This will also allow us to support queries for technology
support/availability.

Fixes: #249