Add functions for querying available technologies
Each plugin (implementation) should be able to report which
technologies it can provide -- which it supports & has all the
required dependencies available for.

This way some plugin implementations may implement only a subset
of the plugin's overall feature set as well as the runtime
feature set may be limitted due to missing runtime dependencies
(usually utilities).