Simplify what WITH_BD_BCACHE changes in the KBD plugin
The primary reason why WITH_BD_BCACHE was introduced was to
prevent the KBD plugin from failing to load when the 'make-bache'
utility was not available. Now that we have a bit more clever
dependency checking, it's easy to just skip this particular check
when loading the plugin when WITH_BD_BCACHE is not defined.

On the other hand, all the bcache code may be available in the
plugin and thus not complicate things terribly. Function calls
would either work or fail based on the availability of their
dependencies.

We need to keep the '--with(out)-bcache' configure options
available for backwards compatibility as well as for tailoring
the spec file.