Add a function to test if prog. reporting was initialized
There was no way to know if libblockdev will report progress outside of
exec.c file. Add a new function to exec.c that will provide this
information. That allows us to modify arguments of executed CLI tools
depending on if they should report progress or not directly in each
plugin's file.

Signed-off-by: Jan Tulak <jtulak@redhat.com>