Discard messages from libdevmapper in the LVM plugins
Otherwise these messages go to the stderr.

This is not an ideal solution, we should log these messages
instead, but that requires a deep analysis of the
dm_log_with_errno_fn callback type and its arguments. If nothing
else at least the log level may need to be translated. See
GH-401.