Add a decorator for "tagging" tests
Tests function (or classes) can be tagged with one or more tags.
This provides easier way to skip slow/unstable/dangerous tests
and/or or some subset of the tests (e.g. run only unstable tests
to check if the test is still failing).

Implemented tags include:
SLOW -- for slow tests (e.g. tests that create LUKS devices)
UNSTABLE -- for tests that fail randomly because the tool/library
            we use is not stable (e.g. some MD RAID or zRAM tests)
UNSAFE -- for tests that change system configuration (e.g. mount
          tests that change /etc/fstab)
CORE -- for tests that cover "core" functionality (this is mostly
        subjective, this should include basic library functions
        and most important test cases for every plugin)
NOSTORAGE -- for tests that don't need storage devices
EXTRADEPS -- for tests with special dependencies (e.g. a special
             hardware like NVDIMM or s390)
REGRESSION -- for regression tests