Move the NTFS read-only device test to a separate test case
And also skip this test case on Debian. On Debian mounting a read
only NTFS device succeeds even if 'rw' option is specified.

On the other hand on Fedora mounting without the 'rw' option fails
instead of mounting the device as read-only. NTFS mounting is just
weird but the test is now written in a way which expects the same
behaviour other filesystems have.

This is not a bug in libblockdev, running mount manually shows the
same inconsistent behaviour.