Fix LVM_MAX_LV_SIZE in the GIR file
GObject instrospection doesn't automatically detect types of the
constants and thinks that all are gint, which doesn't work for
LVM_MAX_LV_SIZE because it's bigger than gint64 and should be
guint64.
Unfortunately there is no way to make the constant in GIR arch
dependent, so we are deprecating it, users shoud use the
bd_lvm_get_max_lv_size function instead.