Do not remove all "flags" in bd_part_set_part_flags on GPT
We want to remove all preexisting flags in this function, but only
"real" flags for given partition type, we don't want to remove or
reset GUIDs on GPT.

This is the good old parted flags vs. real flags. Some of the
flags that parted uses are in fact GUIDs on GPT and we don't want
to touch these, only the real GPT flags which is already covered
by only setting the requested flag(s) when calling "set_gpt_flags"

This is the expected functionality consistent with the previous
parted based implementation, see:

https://github.com/storaged-project/udisks/issues/418
https://github.com/storaged-project/libblockdev/pull/306