Manually add metadata partitions to the list of partitions
When searching partition by position, libparted adds special
metadata partitions at the start of the extended partition and
in front of every logical partition.
libfdisk doesn't do this so we need to add such partitions
manually when creating the list of all partitions for functions
that search for a partition by its position.