fs_test: Fix generic resize for f2fs
We don't need to provide function to check size of a filesystem,
we are using the generic bd_fs_get_size function now.
This was changed in bf0a0b6c1af0acbbdd50494965fe32364642c59a but
the f2fs test were skipped because the technology availability
check didn't work.