fs: Wipe all signatures when running wipe with wipe_all set to TRUE
When wiping filesystem signatures from a device, we are not trying
to wipe all signatures, but only signatures matching the filesystem
name -- so for example all "ntfs" signatures.
This makes sense for FAT with multiple "vfat" signatures, but
NTFS has one "ntfs" signature and one nested "dos" partition table
which we also need to remove.
This change makes sure all signatures (no only filesystem ones and
not only these matching the filesystem name) are wiped from the
device.