tests: Refactor XFS resize tests
We prepare an LVM setup for XFS resize tests and currently if one
the steps of pv/vg/lvcreate fails we don't return error and
continue with the test which makes debugging failed tests really
hard. This change makes sure all errors are raised and moves the
LVM code to a function that can be shared in multiple tests.