tests: Remove test for NTFS read-only mounting
With latest version of ntfs-3g now available in Fedora, mounting
a read only device no longer returns an error, the device is just
mounted as read only with a warning being printed to stdout.
This test was originally added to check that error reporting from
libmount works correnctly with helper programs so it doesn't makes
sense to keep it when mount.ntfs no longer returns an error.