tests: Skip nvme hostnqn tests on non-standard project prefixes
As long as reading the files is done by libnvme and there's no way
to set a different prefix for /etc, let's limit the tests to run
only on matching prefixes, i.e. the standard /usr.

This is done by grepping the configured Makefile and uses undocumented
autotools feature where different prefixes than the standard /usr
are stored and formatted differently in this file.