smart: Remove the ATA low-power mode detection
This never worked as intended and was generally buggy, not being
able to distinguish a real error code. By making the arguments less
restricted this will also allow SAT protocol to be used.

This commit serves as a point of git history in case someone wants
to bring this functionality back.