spec: Bump release to 195 for Packit daily builds
Our custom script we used previously for daily builds bumps the
release number for every build, so the latest libblockdev build
in Copr is 2.99-194 so we need to bump the release to 195 to make
sure the Packit builds will be newer than that. We can reset this
in the next upstream release.