spec: Remove dependency on dosfstools and xfsprogs from fs-devel
This was originally added in d617ba370d2df5dad4a1070aaaf751bba46301c5
but we don't really want to depend on the runtime dependencies and
definitely not in the devel package. Libblockdev users needs to
deal with the dependencies themselves and use the dependency checks
in libblockdev to check for runtime dependencies.