fs: Don't require XFS filesystem to be mounted when getting info
It's not necessary and potentially interferes with other activities on
the system.

(For example, when Cockpit is open while a new Stratis snapshot is
created, there will be a well timed request to UDisks2 that ultimately
mounts the new snapshot filesystem to get its size, right when Stratis
runs xfs_db to change the UUID, which will fail because it is
unexpectedly mounted.  Amazingly narrow race, but it happens.)