fs/xfs: Open device read-only for xfs_info
Running xfs_infor for a device that is not mounted ultimately executes
"xfs_db -c info". Unfortunately, xfs_db opens the device read-write by
default. This results in a udev event.

The udev event in turn might result in another call to xfs_info via
UDisks2 and Cockpit, which conspire the requests the size of a
filesystem for each uevent on it.

Luckily, we can use xfs_db directly and explicitly tell it to open the
device read-only, with the "-r" option.