nvme: Drop the discovery functionality
Dropping for the sole purpose of having clean and stable
libblockdev-3.0 API. To be reverted in later releases.

The nvme discovery works fine at its current form, however
there were multiple changes to the NVMe spec in the meantime
and we need to catch up with the changes. That would inevitably
lead to API break and since there's no public user yet, it's better
to avoid putting any API stability guarantees for a public API
that will be automatically granted by making a public release
of libblockdev.