smart: Introduce well-known attribute names, validation and pretty values
This introduces a level of abstraction for attribute names and
its values. As long as the smartmontools drivedb is vastly
inconsistent in attribute names, a whitelist for each well
known attributes had to be introduced. It serves as
a validation point as well and well known name is provided
only once it passes the whitelist.

Along with attribute name validation, a definition
of attribute value unit is carried in the table. So besides
the well known name a parsed value and its unit is provided
as well.