fs: Add a generic function to check for fs info availability
We don't have a generic function to get filesystem info, but it
would still be usefull to be able to easily query whether the
functionality is available or not. This is similar to the existing
can_get_size and can_get_free_space functions, but these can have
some other limits in them (e.g. XFS not having the free space
information) that can make a difference between "I can query this
FS" and "I can get size/free space of this FS".