smart: Use smartmontools drivedb.h for libatasmart validation
Since we already have a whitelist/conversion table between
smartmontools drivedb.h-style attributes to a libatasmart
well-known attribute names, an idea of additional validation
against drivedb.h came up.

This commit implements a very basic drivedb.h parser for
a built-in include. This needs to be supplied externally
from smartmontools (or downloaded separately for the build).

When enabled, a drive model and firmware is matched against
the drivedb regexes and non-matching attribute definitions
are reported as 'unutrusted'/'unknown'.