btrfs: make btrfs subvolume listing consistent
Depending on the given mountpoint listing subvolumes returns different
pathnames. The pathnames are returned sort of relative to the provided
mountpoint. For example listing with `/` as mountpoint can return:

['root', 'home', 'home/foo', 'foo']

While giving `/home` it returns:

['root', 'home', 'root/foo', 'foo']

This gives issues for Cockpit as we want to give users a consistent
listing of their subvolumes and UDisks always uses the first MountPoint
to look up the subvolumes so Cockpit would either have to figure out how
that is relative to the given MountPoint or show inconsistent
subvolumes.