smart: Rework libatasmart temperature reporting
Libatasmart' temperature_cb() as iterated via sk_disk_smart_get_temperature()
goes through the table and only the last available value
of four matched attributes is reported back to the caller.
This is different from smartmontools which defines matching
priority.

While there are four attributes to be matched, the latter two
are actually overrides from the standard set. This is where
drivedb comes into play, however for the moment we don't have
forward checking. Thus only match #194 and #190, which are
typically safe.