smart: Add BDExtraArg arguments
There's a difference in how the device is accessed. While libatasmart
will happily send requests to the device and just expect an error,
smartctl tries to guess device type and refuses to do anything if unknown
(or abstracted e.g. by dm-multipath). This can be overriden by an extra
argument, libblockdev API just needs a way to provide such override.