tests: Add a separate simpler MD activation test case
MD (de)activation is currently very race-y in both RHEL 9 and 10
and then tests fails often especially when starting and stopping
the array multiple times. This introduces a new simple test case
that should cover most of the use cases without failing all the
time keeping the "full" test case, but marking it as unstable
until these issues are resolved in mdadm.